fastapi-backend/requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database - MySQL with SQLAlchemy
sqlalchemy[asyncio]>=2.0.23
aiomysql>=0.2.0
cryptography>=41.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-decouple>=3.8
email-validator>=2.1.0

# File Handling
python-multipart>=0.0.6
Pillow>=10.0.0

# PDF Generation
reportlab>=4.0.0
weasyprint>=60.0; sys_platform != "win32"  # WeasyPrint works better on Linux/Mac
xhtml2pdf>=0.2.11  # Windows-friendly alternative
qrcode>=7.4.0

# Optional: CORS (included in FastAPI)
# Optional: Redis for caching
redis>=5.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

